pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = 'nomad-lab'
description = 'The NOvel MAterials Discovery (NOMAD) Python package'
readme = "README.md"
authors = [
    { name = "NOMAD Laboratory", email = 'markus.scheidgen@physik.hu-berlin.de' },
]
dynamic = ["version"]
license = { file = "LICENSE" }
requires-python = ">=3.6"
dependencies = [
    'nptyping==1.4.4',
    'cachetools==4.2.4',
    'docstring-parser==0.12',
    'Pint==0.17',
    'orjson==3.6.0',
    'click>=7.1.2',
    'requests==2.27.1',
    'pytz==2021.1',
    'aniso8601==7.0.0',
    'ase==3.19.0',
    'python-keycloak==0.26.1',
    'elasticsearch-dsl==7.4.0',
    'pydantic==1.9.1',
    'jmespath==0.10.0',
    'h5grove>=1.0.0',
    'httpx==0.22.0',
    'memoization==0.4.0',
    'rfc3161ng==2.1.3',
    'lxml~=4.6',
    'wrapt~=1.12.1',
    'pyyaml==6.0'
]

[tool.setuptools_scm]


[project.urls]
homepage = "https://nomad-lab.eu/"
documentation = "https://nomad-lab.eu/prod/v1/docs"
repository = 'https://gitlab.mpcdf.mpg.de/nomad-lab/nomad-FAIR'

[project.optional-dependencies]
parsing = [
    'netCDF4==1.5.4',
    'h5py==3.6.0',
    'hjson==3.0.2',
    'scipy==1.7.1',
    'python-magic==0.4.24',
    'panedr==0.2',
    'parmed==3.0.0',
    'mdtraj==1.9.6',
    'mdanalysis==2.1.0',
    'xarray',
    'phonopy==2.11.0',
    'pymatgen==2022.0.17',
    'asr==0.4.1',
    'bitarray==2.3.5',
    'xrdtools==0.1.1',
    'openpyxl==3.0.9',
    'toposort',
]
infrastructure = [
    'optimade[mongo]==0.18.0',
    'structlog',
    'elasticsearch==7.17.1',
    'msgpack',
    'celery[redis]==4.4.7',
    'asgiref~=3.5',
    'mongoengine>=0.20',
    'pymongo==3.12.1',
    'itsdangerous==2.0.1',
    'passlib==1.7.4',
    'python-logstash==0.4.6',
    'gitpython==3.1.24',
    'm2r==0.2.1',
    'zipstream-new==1.1.5',
    'bagit==1.8.1',
    'bcrypt==3.2.0',
    'filelock==3.3.1',
    'pyjwt[crypto]==1.7.1',
    'cryptography',
    'jsonschema[format]==4.1.2',
    'runstats==2.0.0',
    'tabulate==0.8.9',
    'bs4==0.0.1',
    'html5lib==1.1',
    'zipfile37==0.1.3',
    'basicauth==0.4.1',
    'inflection==0.5.1',
    'unidecode==1.3.2',
    'python-json-logger==2.0.2',
    'recommonmark==0.7.1',
    'jinja2==3.0.3',
    'rdflib==5.0.0',
    'fastapi==0.65.3',
    'uvicorn[standard]',
    'python-multipart==0.0.5',
    'jupyterhub==1.4.2',
    'dockerspawner==12.1.0',
    'oauthenticator==14.2.0',
    'validators==0.18.2',
    'aiofiles==0.8.0',
    'joblib==1.1.0',
    'importlib-metadata~=4.13.0' # Needed because of https://github.com/python/importlib_metadata/issues/411
]
dev = [
    'build',
    'pip-tools',
    'markupsafe',
    'gitpython==3.1.24',
    'mypy==0.730',
    'typed-ast>=1.4.2',
    'astroid>=2.5.1',
    'pylint==2.13.9',
    'pylint_plugin_utils==0.7',
    'pylint_mongoengine==0.4.0',
    'pycodestyle==2.8.0',
    'pytest==3.10.0',
    'pytest-timeout==1.4.2',
    'pytest-cov==2.7.1',
    'rope==0.21.0',
    'names==0.3.0',
    'essential_generators==1.0',
    'twine==3.4.2',
    'python-gitlab==2.10.1',
    'devtools==0.8.0',
    'mkdocs==1.2.3',
    'mkdocs-material==8.1.1',
    'mkdocs-material-extensions==1.0.3',
    'mkdocs-macros-plugin==0.6.3',
]

[project.scripts]
nomad = "nomad.cli:run_cli"

[tool.setuptools.packages.find]
where = [
    ".",
    "dependencies/matid",
    "dependencies/nomad-dos-fingerprints",
    "dependencies/parsers/atomistic",
    "dependencies/parsers/database",
    "dependencies/parsers/eelsdb",
    "dependencies/parsers/electronic",
    "dependencies/parsers/nexus",
    "dependencies/parsers/workflow",
]
exclude = ["tests*"]
namespaces = false
